split sqlIte in smaller files and add test
